+++ modules/frontend/frontend.go
@@ -124,7 +124,7 @@ func NewTracesMiddleware(cfg Config, logger log.Logger, registerer prometheus.Re
 		// - the Deduper dedupes Span IDs for Zipkin support
 		// - the ShardingWare shards queries by splitting the block ID space
 		// - the RetryWare retries requests that have failed (error or http status 500)
-		rt := NewRoundTripper(next, Deduper(logger), ShardingWare(cfg.QueryShards, logger), RetryWare(cfg.MaxRetries, registerer))
+		rt := NewRoundTripper(next, Deduper(logger), ShardingWare(cfg.QueryShards, cfg.TolerateFailedBlocks, logger), RetryWare(cfg.MaxRetries, registerer))
 
 		return RoundTripperFunc(func(r *http.Request) (*http.Response, error) {
 			// validate traceID
@@ -156,15 +156,19 @@ func NewTracesMiddleware(cfg Config, logger log.Logger, registerer prometheus.Re
 				if err != nil {
 					return nil, errors.Wrap(err, "error reading response body at query frontend")
 				}
-				traceObject := &tempopb.Trace{}
-				err = proto.Unmarshal(body, traceObject)
+				responseObject := &tempopb.TraceByIDResponse{}
+				err = proto.Unmarshal(body, responseObject)
 				if err != nil {
 					return nil, err
 				}
 
+				if responseObject.Metrics.FailedBlocks > 0 {
+					resp.StatusCode = http.StatusPartialContent
+				}
+
 				var jsonTrace bytes.Buffer
 				marshaller := &jsonpb.Marshaler{}
-				err = marshaller.Marshal(&jsonTrace, traceObject)
+				err = marshaller.Marshal(&jsonTrace, responseObject.Trace)
 				if err != nil {
 					return nil, err
 				}